requirements.txt
# AI Agent开发团队系统依赖

# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据处理
pydantic==2.5.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1

# AI/LLM框架
langchain==0.1.0
langchain-community==0.0.10
langgraph==0.0.20
autogen-agentchat==0.2.0

# HTTP客户端
httpx==0.25.2
aiohttp==3.9.1

# 向量数据库和Embedding模型
qdrant-client==1.7.0
FlagEmbedding==1.2.5
sentence-transformers==2.2.2
torch>=2.0.0
transformers>=4.33.0

# 日期时间处理
python-dateutil==2.8.2

# 环境变量
python-dotenv==1.0.0

# JSON处理
orjson==3.9.10

# 异步任务
asyncio-mqtt==0.13.0

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# 性能监控
prometheus-client==0.19.0

# 安全
cryptography==41.0.8